Manual Robot for LA Regional - Prevalidation
Updated manual robot code to prepare for LA regional.  This version of
the code needs to be tested and updated for 2016 robot configuration but
it is almost everything needed for the LA regional for a manual robot.
Updated this code to reflect changes recently made to the stated robot
code (2016_Robot branch): Updated parameter configuration (e.g. channel
#s), simplified autonomous code, updated autonomous code to use separate
speeds for left and right sides, added code so network tables will work
as expected